Rename Dataset_TAI-Index_Braun24.xlsx to Data set_TAI-Index_Braun24.xlsx
A data set coding 4 representative transnational informational practices for 178 countries (4278 total obersavations). The data set is built on the armature of V-DEM coding and therewith cross-compatible. 

Besides the combined dataset, the respective codings for Presonalised data localisation practice and International state-owned news broadcasting stations are also contained in the individual Excel-sheets, including the codebooks for the coding.